Update example.rst
I add the last ampl_list numbers, and now we are using False for the centre and sigma in the strongest emission lines than copy Halpha and [OIII]5007, e.g [NII]6548,6584, [OIII]4959 and both lines of [SII]6717,6731. 
I have left only [SII]6717 as example when we use the centre and sigma of Halpha and [OIII]5007 as limits.

Take care, that this is an example when we fit in velocities range. We can add another example when we fit using wavelength ranges.